requirements.txt
# Core ML Framework
mlx>=0.18.0
mlx-lm>=0.0.12

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0
huggingface-hub>=0.19.0
transformers>=4.35.0
tokenizers>=0.15.0

# SQL Processing
sqlparse>=0.4.4
sqlalchemy>=2.0.0

# Training & Optimization
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0

# Evaluation
scikit-learn>=1.3.0
nltk>=3.8.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# Optional: Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
